[fix warnings]::compare_different_sgn results of this commit is different to the previous.

+++ src/gap_util.cpp
@@ -764,10 +764,10 @@ int dropChainGapX(String<uint64_t> & chains,
         for (int i = 1; i < (int)length(chains); i++)
         {
             uint di = i + 1 >= gap_parms.thd_dcgx_window_size ? gap_parms.thd_dcgx_window_size : 1;
-            if (getX(chains[i]) - getX(chains[i - 1]) > gap_parms.thd_dcgx_Xdrop_peak || 
-                getX(chains[i]) - getX(chains[i + 1 - di]) > gap_parms.thd_dcgx_Xdrop_sum ||
-                getY(chains[i]) - getY(chains[i - 1]) > gap_parms.thd_dcgx_Xdrop_peak || 
-                getY(chains[i]) - getY(chains[i + 1 - di]) > gap_parms.thd_dcgx_Xdrop_sum)
+            if (int64_t(getX(chains[i]) - getX(chains[i - 1])) > gap_parms.thd_dcgx_Xdrop_peak ||
+                int64_t(getX(chains[i]) - getX(chains[i + 1 - di])) > gap_parms.thd_dcgx_Xdrop_sum ||
+                int64_t(getY(chains[i]) - getY(chains[i - 1])) > gap_parms.thd_dcgx_Xdrop_peak ||
+                int64_t(getY(chains[i]) - getY(chains[i + 1 - di])) > gap_parms.thd_dcgx_Xdrop_sum)
             {
                 if (f_erase)
                 {
@@ -783,11 +783,11 @@ int dropChainGapX(String<uint64_t> & chains,
     {
         for (int i = length(chains) - 2; i > 0; i--) 
         {
-            uint di = length(chains) - i >= gap_parms.thd_dcgx_window_size ? gap_parms.thd_dcgx_window_size : 1;
-            if (getX(chains[i + 1]) - getX(chains[i]) > gap_parms.thd_dcgx_Xdrop_peak || 
-                getX(chains[i + di - 1]) - getX(chains[i]) > gap_parms.thd_dcgx_Xdrop_sum ||
-                getY(chains[i + 1]) - getY(chains[i]) > gap_parms.thd_dcgx_Xdrop_peak || 
-                getY(chains[i + di - 1]) - getY(chains[i]) > gap_parms.thd_dcgx_Xdrop_sum)
+            uint di = (int)length(chains) - i >= gap_parms.thd_dcgx_window_size ? gap_parms.thd_dcgx_window_size : 1;
+            if (int64_t(getX(chains[i + 1]) - getX(chains[i])) > gap_parms.thd_dcgx_Xdrop_peak ||
+                int64_t(getX(chains[i + di - 1]) - getX(chains[i])) > gap_parms.thd_dcgx_Xdrop_sum ||
+                int64_t(getY(chains[i + 1]) - getY(chains[i])) > gap_parms.thd_dcgx_Xdrop_peak ||
+                int64_t(getY(chains[i + di - 1]) - getY(chains[i])) > gap_parms.thd_dcgx_Xdrop_sum)
             {
                 if (f_erase)
                 {
